Cambios a la estructura, Equpación de Flyway Al proyecto, entidad productos iniciada
