+++ demo/src/main/java/com/inventario/demo/config/ModelMapperConfig.java
@@ -1,6 +1,6 @@
 package com.inventario.demo.config;
 
-import com.inventario.demo.user.mapper.UserMapper;
+import com.inventario.demo.entities.user.mapper.UserMapper;
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
