+++ demo/src/main/java/com/inventario/demo/entities/productos/dtoResponse/ResponseCreateProductDto.java
@@ -1,0 +1,12 @@
+package com.inventario.demo.entities.productos.dtoResponse;
+
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+
+@JsonPropertyOrder({"Nombre","Categoria","SKU"})
+public record ResponseCreateProductDto(
+        String Nombre,
+        String Categoria,
+        String SKU
+
+) {}
