+++ demo/src/main/java/com/inventario/demo/entities/user/Enum/EnumPermission.java
@@ -1,0 +1,8 @@
+package com.inventario.demo.entities.user.Enum;
+
+public enum EnumPermission {
+    CREATE,
+    READ,
+    UPDATE,
+    DELETE
+}
