+++ demo/src/main/java/com/inventario/demo/entities/user/Enum/EnumRole.java
@@ -1,0 +1,6 @@
+package com.inventario.demo.entities.user.Enum;
+
+public enum EnumRole {
+    ADMIN,
+    USER
+}
