+++ demo/src/main/java/com/inventario/demo/entities/user/dtoResponse/AuthResponseDto.java
@@ -1,0 +1,4 @@
+package com.inventario.demo.entities.user.dtoResponse;
+
+public record AuthResponseDto (Long id, String email, String message, String token, boolean success){
+}
