+++ demo/src/main/java/com/inventario/demo/entities/user/dtoResponse/UserPageResponse.java
@@ -1,0 +1,8 @@
+package com.inventario.demo.entities.user.dtoResponse;
+
+import java.util.List;
+
+public record UserPageResponse (List<UserResponseDto> users,
+                                int totalPages,
+                                long totalElements){
+}
