+++ demo/src/main/java/com/inventario/demo/entities/user/model/PermissionModel.java
@@ -1,0 +1,21 @@
+package com.inventario.demo.entities.user.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+
+@Entity
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Table(name = "permission")
+public class PermissionModel {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true, nullable = false, updatable = false)
+    private String name;
+}
