+++ demo/src/main/java/com/inventario/demo/entities/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.inventario.demo.entities.user.repository;
+
+import com.inventario.demo.entities.user.model.UserModel;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<UserModel, Long> {
+    Optional<UserModel> findByEmail(String email);
+}
