--- demo/src/main/java/com/inventario/demo/user/Enum/EnumPermission.java
@@ -1,8 +1,0 @@
-package com.inventario.demo.user.Enum;
-
-public enum EnumPermission {
-    CREATE,
-    READ,
-    UPDATE,
-    DELETE
-}
