--- demo/src/main/java/com/inventario/demo/user/Enum/EnumRole.java
@@ -1,6 +1,0 @@
-package com.inventario.demo.user.Enum;
-
-public enum EnumRole {
-    ADMIN,
-    USER
-}
