--- demo/src/main/java/com/inventario/demo/user/dtoRequest/AuthCreateRoleRequestDto.java
@@ -1,8 +1,0 @@
-package com.inventario.demo.user.dtoRequest;
-
-import jakarta.validation.constraints.Size;
-
-import java.util.List;
-
-public record AuthCreateRoleRequestDto (@Size(max = 1, message = "roles must be a maximum of 1 character in length" ) List<String> roles){
-}
