--- demo/src/main/java/com/inventario/demo/user/dtoResponse/AuthResponseDto.java
@@ -1,4 +1,0 @@
-package com.inventario.demo.user.dtoResponse;
-
-public record AuthResponseDto (Long id, String email, String message, String token, boolean success){
-}
