--- demo/src/main/java/com/inventario/demo/user/dtoResponse/AuthResponseRegisterDto.java~
@@ -1,8 +1,0 @@
-package com.inventario.demo.user.dtoResponse;
-
-public record AuthResponseRegisterDto (Long userId,
-                                      String email,
-                                      String message,
-                                      String jwt,
-                                      boolean status){
-}
