--- demo/src/main/java/com/inventario/demo/user/dtoResponse/UserPageResponse.java
@@ -1,10 +1,0 @@
-package com.inventario.demo.user.dtoResponse;
-
-import com.inventario.demo.user.dtoRequest.UserRequestDto;
-
-import java.util.List;
-
-public record UserPageResponse (List<UserResponseDto> users,
-                                int totalPages,
-                                long totalElements){
-}
