--- demo/src/main/java/com/inventario/demo/user/dtoResponse/UserResponseDto.java
@@ -1,23 +1,0 @@
-package com.inventario.demo.user.dtoResponse;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDate;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserResponseDto {
-    private Long id;
-    private String email;
-    private String name;
-    private String lastName;
-    private Long phoneNumber;
-    private LocalDate birthDate;
-    private LocalDate registerDate;
-    private LocalDate lastLogin;
-}
