--- demo/src/main/java/com/inventario/demo/user/model/PermissionModel.java~
@@ -1,15 +1,0 @@
-package com.inventario.demo.user.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-public class PermissionModel {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true, nullable = false, updatable = false)
-    private String name;
-}
