--- demo/src/main/java/com/inventario/demo/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.inventario.demo.user.repository;
-
-import com.inventario.demo.user.model.UserModel;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserModel, Long> {
-    Optional<UserModel> findByEmail(String email);
-}
